+++ CHANGELOG.md
@@ -1,3 +1,15 @@
 ## 0.0.1
 
 * Initial release of the Smart Edge Alert package.
+
+## 0.0.2
+
+* Updates documentation for Class and Method descriptions.
+
+## 0.0.3
+
+* Updates README.md with a detailed description of how to use the Smart Edge Alert.
+
+## 0.0.4
+
+* Adds titleFontSize and descriptionFontSize properties to the SmartEdgeAlert.show method to allow for content customization.